iox-#1032 Polymorphic Singleton
