+++ iceoryx_hoofs/include/iceoryx_hoofs/error_handling_2/platform/test_platform/handler_interface.hpp
@@ -4,9 +4,6 @@
 #include "iceoryx_hoofs/error_handling_2/location.hpp"
 #include "iceoryx_hoofs/error_handling_2/platform/error_levels.hpp"
 
-#include <atomic>
-#include <mutex>
-
 namespace eh
 {
 // can be defined by the platform and will/should depend on the error levels of the platform
@@ -21,69 +18,5 @@ struct HandlerInterface
 
     virtual void preterminate() = 0;
 };
-
-// infrastructure to have exactly one handler installed
-
-// provides the currently installed unique handler, initially default
-template <typename Interface, typename DefaultHandler>
-class UniqueHandler
-{
-  public:
-    static Interface& get()
-    {
-        return *s_handler.load(std::memory_order_relaxed);
-    }
-
-    // we always require a valid handler to be set
-    // enforce contract of valid handler
-    static Interface* set(Interface& handler)
-    {
-        auto& h = instance();
-        std::lock_guard<std::mutex> lock(h.m_mutex);
-        if (h.m_isFinal)
-        {
-            std::cerr << "SETTING THE ERROR HANDLER AFTER FIANALIZE IS NOT ALLOWED!" << std::endl;
-            std::terminate();
-            return nullptr;
-        }
-
-        return s_handler.exchange(&handler, std::memory_order_relaxed);
-    }
-
-    static Interface* reset()
-    {
-        return set(s_default);
-    }
-
-    static void finalize()
-    {
-        auto& h = instance();
-        std::lock_guard<std::mutex> lock(h.m_mutex);
-        h.m_isFinal.store(true);
-    }
-
-  private:
-    static DefaultHandler s_default; // ensure there is always at least a default handler
-    // not a member to avoid indirection - we may have inidrection ininstance anyway
-    static std::atomic<Interface*> s_handler;
-
-    // avoid lots of static variables, these are not accessed often (set and finalize are rare)
-    std::atomic_bool m_isFinal{false};
-    std::mutex m_mutex; // required for sync only
-
-    // there is only ever one of them but it cannot be accessed
-    static UniqueHandler& instance()
-    {
-        static UniqueHandler h;
-        return h;
-    }
-};
-
-// these are initialized before any static method (set, get etc.) is called(!)
-template <typename I, typename D>
-D UniqueHandler<I, D>::s_default{};
-
-template <typename I, typename D>
-std::atomic<I*> UniqueHandler<I, D>::s_handler{&s_default};
 
 } // namespace eh
